requirements_pip.txt
psycopg2-binary
django-prettyjson
django-redis
pytest-django
codecov
django-debug-toolbar
pdfkit
sklearn
xxhash
black
responder
cherrypy
django-storages
awscli
sentry-sdk
graphviz
pytest-profiling<1.4.0
pysnooper

# 맥에서는 설치 실패함. 허나 uri_hash 계산시에만 사용되고 맥에서는 xxhash 로 대체하도록 코드가 짜여져 있어 문제 없음
cityhash
